+++ src/index.js
@@ -4,7 +4,10 @@
  * 工具库、webpack基础配置、bid工具Express中间件
  * 未经本人允许禁止商业用途
  */
-console.log('bid-base version:', require('../package.json').version)
+const { name, version } = require('../package.json')
+
+console.log(`${name} version:${version}`)
+
 export { default as Utils } from './utils/index'
 export { default as getIp } from './utils/getip'
 export { default as Lint } from './actions/lint'
@@ -17,4 +20,4 @@ export { default as CompareVersion } from './utils/compare-version'
 export { default as WebpackProConfig } from './webpack/webpack.pro.config'
 export { default as WebpackDevConfig } from './webpack/webpack.dev.config'
 export { default as WebpackTestConfig } from './webpack/webpack.test.config'
-export { merge as WebpackMerge } from 'webpack-merge'
+export { merge as WebpackMerge, mergeWithCustomize, mergeWithRules, unique } from 'webpack-merge'