+++ src/utils/get-alias.js
@@ -1,4 +1,4 @@
-import path from 'path';
+import path from 'path'
 
 const getAlias = userAlias => {
     /*
@@ -6,14 +6,15 @@ const getAlias = userAlias => {
      */
     let alias = {
         // 别名
-        '@br': path.join(process.cwd(), './src/c/')
-    };
-    // eslint-disable-next-line
-    for (let aliasName in userAlias) {
-        userAlias[aliasName] = userAlias[aliasName].replace(/^@br\//, alias['@br']); // 将别名配置中的@br替换为'src/c'目录
+        '@br': path.join(process.cwd(), './src/c/'),
+        '@src': path.join(process.cwd(), './src/')
     }
-    alias = Object.assign(alias, userAlias);
-    return alias;
-};
+    // eslint-disable-next-line
+    // for (let aliasName in userAlias) {
+    //     userAlias[aliasName] = userAlias[aliasName].replace(/^@br\//, alias['@br']).replace(/^@src\//, alias['@src']) // 将别名配置中的@br替换为'src/c'目录
+    // }
+    alias = Object.assign(alias, userAlias)
+    return alias
+}
 
-export default getAlias;
+export default getAlias