+++ src/webpack/plugins.js
@@ -4,12 +4,8 @@ import webpack from 'webpack'
 import Utils from '../utils'
 import HtmlWebpackPlugin from '../plugins/html-webpack-plugin'
 import htmlInlineSourceLoaders from '../plugins/html-inline-source-loaders'
-// import { CleanWebpackPlugin } from 'clean-webpack-plugin'
 import HtmlWebpackReplaceHost from 'html-webpack-replace-host'
-// import LazyPathPlugin from 'bid-lazy-path-plugin'
 import HtmlWebpackInlineSourcePlugin from 'webpack-plugin-inline-source'
-// import FriendlyErrorsPlugin from 'friendly-errors-webpack-plugin'
-// import ProgressBar from 'progress-bar-webpack-plugin'
 
 const tagVersion = Utils.getUserConfig.version
 
@@ -23,31 +19,33 @@ function setHtmlPlugin(file, env) {
             collapseWhitespace: true,
             removeComments: true
         },
+        meta: {
+            env: env,
+            'update-time': { 'update-time': new Date().toLocaleString() }
+        },
         path: path.resolve(process.cwd(), './deploy', isOnline ? './html/build' : './build'),
         filename: file,
         template: path.resolve(file)
     })
 }
 
-function getPlugins({ htmlEntry, env = 'local', cdnhost }) {
+function getPlugins({ htmlEntry, env = 'daily', cdnhost }) {
     const jsHost = `${cdnhost || Utils.getUserConfig.cdnhost}/${Utils.getUserConfig.appName}/`
     const jsPath = jsHost + path.dirname(`${htmlEntry.replace(/^\.\//, '')}`) + `/${Utils.getUserConfig.version}/`
 
     let config = [
-        new webpack.DefinePlugin({ NODE_ENV: JSON.stringify(process.env.NODE_ENV) }),
+        new webpack.DefinePlugin({
+            NODE_ENV: JSON.stringify(process.env.NODE_ENV),
+            BUILD_ENV: JSON.stringify(env)
+        }),
         //避免重复的模块
-        // new webpack.optimize.DedupePlugin()
+        new webpack.optimize.DedupePlugin(),
         /* 跳过编译时出错的代码并记录 , webpack.NoErrorsPlugin webpack4后改为webpack.NoEmitOnErrorsPlugin */
         new webpack.NoEmitOnErrorsPlugin()
     ]
 
-    if (process.env.NODE_ENV != 'dev') {
-        if (env == 'local' || env == 'daily' || env == 'production-build') {
-            // config.unshift(new ProgressBar())
-            // config.push(new CleanWebpackPlugin())
-            config.push(new webpack.ProgressPlugin({ percentBy: 'entries' }))
-        }
-
+    if (env == 'local' || env == 'daily' || env == 'production-build') {
+        config.push(new webpack.ProgressPlugin({ percentBy: 'entries' }))
         // config.push(new LazyPathPlugin({ version: tagVersion, jsHost, env }))
         config.push(setHtmlPlugin(htmlEntry, env))